django_projects/todo_api/requirements.txt
Pillow
django
djangorestframework
psycopg2-binary